Configure kmp related logic in `ShadowKmpPlugin`
